Divide into multiple packages
