+++ src/main/java/zoe/ZoeException.java
@@ -1,3 +1,5 @@
+package zoe;
+
 public class ZoeException extends Exception {
     public ZoeException(String message) {
         super(message);
