+++ src/main/java/zoe/ZoeIOException.java
@@ -1,3 +1,5 @@
+package zoe;
+
 public class ZoeIOException extends ZoeException {
     public ZoeIOException(String message) {
         super(message);
