+++ src/main/java/zoe/adapters/gson/LocalDateTypeAdapter.java
@@ -1,3 +1,5 @@
+package zoe.adapters.gson;
+
 import com.google.gson.*;
 
 import java.lang.reflect.Type;
