+++ src/main/java/zoe/command/DeadlineCommand.java
@@ -1,4 +1,12 @@
+package zoe.command;
+
 import lombok.AllArgsConstructor;
+import zoe.ZoeException;
+import zoe.storage.Storage;
+import zoe.task.Deadline;
+import zoe.task.Task;
+import zoe.task.TaskList;
+import zoe.ui.Ui;
 
 import java.time.LocalDate;
