+++ src/main/java/zoe/command/IncorrectCommand.java
@@ -1,5 +1,11 @@
+package zoe.command;
+
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
+import zoe.ZoeException;
+import zoe.storage.Storage;
+import zoe.task.TaskList;
+import zoe.ui.Ui;
 
 @NoArgsConstructor
 @AllArgsConstructor
