+++ src/main/java/zoe/task/Deadline.java
@@ -1,3 +1,5 @@
+package zoe.task;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
@@ -6,19 +8,18 @@ import java.time.format.DateTimeFormatter;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
-public class Event extends Task {
+public class Deadline extends Task {
     private final String taskName;
-    private final LocalDate from;
-    private final LocalDate to;
+    private final LocalDate by;
 
     @Override
     public String getTypeIcon() {
-        return "E";
+        return "D";
     }
 
     @Override
     public String getDescription() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM d yyyy");
-        return String.format("%s (from: %s to: %s)", taskName, from.format(formatter), to.format(formatter));
+        return String.format("%s (by: %s)", taskName, by.format(formatter));
     }
 }
