+++ src/main/java/zoe/task/Task.java
@@ -1,3 +1,5 @@
+package zoe.task;
+
 import lombok.Data;
 
 @Data
