+++ src/main/java/zoe/ui/Ui.java
@@ -1,4 +1,9 @@
+package zoe.ui;
+
 import lombok.NoArgsConstructor;
+import zoe.ZoeException;
+import zoe.task.Task;
+import zoe.task.TaskList;
 
 import java.util.ListIterator;
 import java.util.Scanner;
@@ -11,7 +16,7 @@ public class Ui {
     }
 
     public void showWelcome() {
-        System.out.println("Hello! I'm Zoe\nWhat can I do for you?\n");
+        System.out.println("Hello! I'm zoe.Zoe\nWhat can I do for you?\n");
     }
 
     public void showGoodbye() {
