+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Zoe
+Hello! I'm zoe.Zoe
 What can I do for you?
 
 _______
